+++ java/src/com/presnakov/hotelBookingSystem/dto/room/RoomClassDto.java
@@ -4,12 +4,10 @@ import com.presnakov.hotelBookingSystem.entity.RoomClassEnum;
 import lombok.Builder;
 import lombok.Value;
 
-import java.math.BigDecimal;
-
 @Value
 @Builder
 public class RoomClassDto {
     Integer id;
     RoomClassEnum comfortClass;
-    BigDecimal pricePerDay;
+    Double pricePerDay;
 }
